Shanoir-execution -> change logic
